requirements.txt
azure-functions
openai
azure-monitor-opentelemetry
azure-identity
